create version checker util and use it in auto-update, display-banner
