+++ packages/cli/src/utils/index.ts
@@ -22,3 +22,4 @@ export * from './resolve-utils';
 export * from './run-bun';
 export * from './test-runner';
 export * from './user-environment';
+export * from './version';
